Backend/src/main/java/com/example/Kanban/KanbanBoard/KanbanBoardController.java
package com.example.Kanban.KanbanBoard;

import com.example.Kanban.KanbanBoard.dto.CreateKanbanBoardDTO;
import com.example.Kanban.KanbanBoard.dto.KanbanBoardResponse;
import lombok.AllArgsConstructor;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/boards")
@AllArgsConstructor
public class KanbanBoardController {

    private final KanbanBoardService boardService;

    @PostMapping("/create")
    public ResponseEntity<KanbanBoardResponse> createKanbanBoard(CreateKanbanBoardDTO createKanbanBoardDTO) {
        return boardService.createKanbanBoard(createKanbanBoardDTO);
    }
}
